Embed views in a tabView
