+++ Ponto-MVVM/Ponto-MVVM/App/Ponto_MVVMApp.swift
@@ -11,7 +11,18 @@ import SwiftUI
 struct Ponto_MVVMApp: App {
     var body: some Scene {
         WindowGroup {
-            TimeCardView(viewModel: CurrentTimeCardViewModel())
+            TabView {
+                EmbeddedViewInNavigation(embeddedView: AnyView(TimeCardView(viewModel: CurrentTimeCardViewModel())))
+                    .tabItem {
+                        Image(systemName: Constants.ImageName.calendarIcon)
+                        Text(Constants.CurrentTimeCard.tabBarTitle)
+                    }
+                EmbeddedViewInNavigation(embeddedView: AnyView(TimeCardHistoryView(viewModel: TimeCardHistoryViewModel())))
+                    .tabItem {
+                        Image(systemName: Constants.ImageName.clockIcon)
+                        Text(Constants.TimeCardHistory.screenTitle)
+                    }
+            }
         }
     }
 }
