+++ Ponto-MVVM/Ponto-MVVM/View/EmbeddedViewInNavigation.swift
@@ -1,0 +1,20 @@
+//
+//  EmbeddedViewInNavigation.swift
+//  Ponto-MVVM
+//
+//  Created by Carolina Cruz Agra Lopes on 31/08/22.
+//
+
+import SwiftUI
+
+struct EmbeddedViewInNavigation: View {
+
+    let embeddedView: AnyView
+
+    var body: some View {
+        NavigationView {
+            embeddedView
+        }
+    }
+
+}
